Created a new module vip-publication out of the existing publication in vip-code
