+++ vip-core/src/main/java/fr/insalyon/creatis/vip/core/client/view/main/GeneralTileGrid.java
@@ -38,7 +38,6 @@ import fr.insalyon.creatis.vip.core.client.view.application.ApplicationTileRecor
 import fr.insalyon.creatis.vip.core.client.view.application.ApplicationsTileGrid;
 import fr.insalyon.creatis.vip.core.client.view.layout.Layout;
 import fr.insalyon.creatis.vip.core.client.view.user.AccountTab;
-import fr.insalyon.creatis.vip.core.client.view.user.PublicationTab;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -54,7 +53,6 @@ public class GeneralTileGrid extends ApplicationsTileGrid {
 
         super("General");
         addApplication(CoreConstants.APP_ACCOUNT, CoreConstants.APP_IMG_ACCOUNT);
-        addApplication(CoreConstants.APP_PUBLICATIONS, CoreConstants.APP_IMG_PUBLICATIONS);
         parsers = new ArrayList<ApplicationParser>();
     }
 
@@ -64,12 +62,6 @@ public class GeneralTileGrid extends ApplicationsTileGrid {
         if (applicationName.equals(CoreConstants.APP_ACCOUNT)) {
             Layout.getInstance().addTab(
                 CoreConstants.TAB_ACCOUNT, AccountTab::new);
-            return;
-        }
-
-        if (applicationName.equals(CoreConstants.APP_PUBLICATIONS)) {
-            Layout.getInstance().addTab(
-                CoreConstants.TAB_PUBLICATION, PublicationTab::new);
             return;
         }
