+++ vip-portal/src/main/java/fr/insalyon/creatis/vip/portal/Main.gwt.xml
@@ -45,6 +45,7 @@ knowledge of the CeCILL-B license and that you accept its terms.
     <inherits name="fr.insalyon.creatis.vip.social.Social" />
     <inherits name="fr.insalyon.creatis.vip.applicationimporter.ApplicationImporter" />
     <inherits name="fr.insalyon.creatis.vip.visualization.Visualization" />
+    <inherits name="fr.insalyon.creatis.vip.publication.Publication"/>
 
     <!-- Specify the app entry point class.-->
     <entry-point class="fr.insalyon.creatis.vip.portal.client.Main"/>
