+++ vip-publication/src/main/java/fr/insalyon/creatis/vip/publication/client/rpc/PublicationService.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright and authors: see LICENSE.txt in base repository.
+ *
+ * This software is a web portal for pipeline execution on distributed systems.
+ *
+ * This software is governed by the CeCILL-B license under French law and
+ * abiding by the rules of distribution of free software.  You can  use, 
+ * modify and/ or redistribute the software under the terms of the CeCILL-B
+ * license as circulated by CEA, CNRS and INRIA at the following URL
+ * "http://www.cecill.info". 
+ *
+ * As a counterpart to the access to the source code and  rights to copy,
+ * modify and redistribute granted by the license, users are provided only
+ * with a limited warranty  and the software's author,  the holder of the
+ * economic rights,  and the successive licensors  have only  limited
+ * liability. 
+ *
+ * In this respect, the user's attention is drawn to the risks associated
+ * with loading,  using,  modifying and/or developing or reproducing the
+ * software by the user in light of its specific status of free software,
+ * that may mean  that it is complicated to manipulate,  and  that  also
+ * therefore means  that it is reserved for developers  and  experienced
+ * professionals having in-depth computer knowledge. Users are therefore
+ * encouraged to load and test the software's suitability as regards their
+ * requirements in conditions enabling the security of their systems and/or 
+ * data to be ensured and,  more generally, to use and operate it in the 
+ * same conditions as regards security. 
+ *
+ * The fact that you are presently reading this means that you have had
+ * knowledge of the CeCILL-B license and that you accept its terms.
+ */
+package fr.insalyon.creatis.vip.publication.client.rpc;
+
+import com.google.gwt.core.client.GWT;
+import com.google.gwt.user.client.rpc.RemoteService;
+import com.google.gwt.user.client.rpc.ServiceDefTarget;
+import fr.insalyon.creatis.vip.publication.client.bean.Publication;
+import fr.insalyon.creatis.vip.core.client.view.CoreException;
+import java.util.List;
+
+/**
+ *
+ * @author Rafael Ferreira da Silva, Sorina Pop
+ */
+public interface PublicationService extends RemoteService {
+
+    String SERVICE_URI = "/publicationservice";
+
+    class Util {
+
+        public static PublicationServiceAsync getInstance() {
+
+            PublicationServiceAsync instance = GWT.create(PublicationService.class);
+            ServiceDefTarget target = (ServiceDefTarget) instance;
+            target.setServiceEntryPoint(GWT.getModuleBaseURL() + SERVICE_URI);
+            return instance;
+        }
+    }
+
+    void addPublication(Publication pub) throws CoreException;
+
+    void updatePublication(Publication pub) throws CoreException;
+
+    List<Publication> getPublications() throws CoreException;
+
+    void removePublication(Long id) throws CoreException;
+
+    List<Publication> parseBibtexText(String text) throws CoreException;
+
+}
