+++ vip-publication/src/main/java/fr/insalyon/creatis/vip/publication/client/view/PublicationGrid.java
@@ -29,7 +29,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-B license and that you accept its terms.
  */
-package fr.insalyon.creatis.vip.core.client.view.user.publication;
+package fr.insalyon.creatis.vip.publication.client.view;
 
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.smartgwt.client.data.DataSource;
@@ -51,11 +51,10 @@ import com.smartgwt.client.widgets.layout.HLayout;
 import com.smartgwt.client.widgets.viewer.DetailViewer;
 import com.smartgwt.client.widgets.viewer.DetailViewerField;
 import fr.insalyon.creatis.vip.core.client.CoreModule;
-import fr.insalyon.creatis.vip.core.client.rpc.ConfigurationService;
 import fr.insalyon.creatis.vip.core.client.view.CoreConstants;
 import fr.insalyon.creatis.vip.core.client.view.layout.Layout;
-import fr.insalyon.creatis.vip.core.client.view.user.PublicationTab;
 import fr.insalyon.creatis.vip.core.client.view.user.UserLevel;
+import fr.insalyon.creatis.vip.publication.client.rpc.PublicationService;
 
 /**
  *
@@ -209,7 +208,7 @@ public class PublicationGrid extends ListGrid {
     protected static void edit(String id, String title, String type, String typeName, String authors, String date, String doi, String vipApplication) {
 
         PublicationTab pubTab = (PublicationTab) Layout.getInstance().
-                getTab(CoreConstants.TAB_PUBLICATION);
+                getTab(PublicationConstants.TAB_PUBLICATION);
         pubTab.setPublication(id, title, type, typeName, authors, date, doi, vipApplication);
     }
 
@@ -232,7 +231,7 @@ public class PublicationGrid extends ListGrid {
         };
        getPubTab().getModal().show("Removing publication '" + "'...", true);
 
-        ConfigurationService.Util.getInstance().removePublication(id, callback);
+        PublicationService.Util.getInstance().removePublication(id, callback);
 
     }
 
@@ -251,6 +250,6 @@ public class PublicationGrid extends ListGrid {
 
     private PublicationTab getPubTab() {
         return (PublicationTab) Layout.getInstance().
-                getTab(CoreConstants.TAB_PUBLICATION);
+                getTab(PublicationConstants.TAB_PUBLICATION);
     }
 }
