+++ vip-publication/src/main/java/fr/insalyon/creatis/vip/publication/server/dao/MySQLDAOFactory.java
@@ -4,16 +4,16 @@
  * This software is a web portal for pipeline execution on distributed systems.
  *
  * This software is governed by the CeCILL-B license under French law and
- * abiding by the rules of distribution of free software.  You can  use, 
+ * abiding by the rules of distribution of free software.  You can  use,
  * modify and/ or redistribute the software under the terms of the CeCILL-B
  * license as circulated by CEA, CNRS and INRIA at the following URL
- * "http://www.cecill.info". 
+ * "http://www.cecill.info".
  *
  * As a counterpart to the access to the source code and  rights to copy,
  * modify and redistribute granted by the license, users are provided only
  * with a limited warranty  and the software's author,  the holder of the
  * economic rights,  and the successive licensors  have only  limited
- * liability. 
+ * liability.
  *
  * In this respect, the user's attention is drawn to the risks associated
  * with loading,  using,  modifying and/or developing or reproducing the
@@ -22,33 +22,38 @@
  * therefore means  that it is reserved for developers  and  experienced
  * professionals having in-depth computer knowledge. Users are therefore
  * encouraged to load and test the software's suitability as regards their
- * requirements in conditions enabling the security of their systems and/or 
- * data to be ensured and,  more generally, to use and operate it in the 
- * same conditions as regards security. 
+ * requirements in conditions enabling the security of their systems and/or
+ * data to be ensured and,  more generally, to use and operate it in the
+ * same conditions as regards security.
  *
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-B license and that you accept its terms.
  */
-package fr.insalyon.creatis.vip.core.client.view.user.publication;
+package fr.insalyon.creatis.vip.publication.server.dao;
 
-import com.smartgwt.client.widgets.grid.ListGridRecord;
+import fr.insalyon.creatis.vip.core.server.dao.DAOException;
+import fr.insalyon.creatis.vip.publication.server.dao.mysql.PublicationData;
+
+import java.sql.Connection;
 
 /**
  *
- * @author Nouha Boujelben
+ * @author Rafael Ferreira da Silva,Nouha boujelben
  */
-public class PublicationRecord extends ListGridRecord {
-
-    public PublicationRecord(Long id, String title, String type, String typeName, String date, String authors, String doi, String vipAuthor, String vipApplication) {
-
-        setAttribute("id", id);
-        setAttribute("title", title);
-        setAttribute("type", type);
-        setAttribute("typeName", typeName);
-        setAttribute("date", date);
-        setAttribute("authors", authors);
-        setAttribute("doi", doi);
-        setAttribute("vipAuthor", vipAuthor);
-        setAttribute("vipApplication", vipApplication);
+public class MySQLDAOFactory extends PublicationDAOFactory {
+    private static PublicationDAOFactory instance;
+    // Singleton
+    protected static PublicationDAOFactory getInstance() {
+        if (instance == null) {
+            instance = new MySQLDAOFactory();
+        }
+        return instance;
     }
+
+    @Override
+    public PublicationDAO getPublicationDAO(Connection connection)
+        throws DAOException {
+        return new PublicationData(connection);
+    }
+
 }
