+++ vip-publication/src/main/java/fr/insalyon/creatis/vip/publication/server/dao/PublicationDAO.java
@@ -4,16 +4,16 @@
  * This software is a web portal for pipeline execution on distributed systems.
  *
  * This software is governed by the CeCILL-B license under French law and
- * abiding by the rules of distribution of free software.  You can  use,
+ * abiding by the rules of distribution of free software.  You can  use, 
  * modify and/ or redistribute the software under the terms of the CeCILL-B
  * license as circulated by CEA, CNRS and INRIA at the following URL
- * "http://www.cecill.info".
+ * "http://www.cecill.info". 
  *
  * As a counterpart to the access to the source code and  rights to copy,
  * modify and redistribute granted by the license, users are provided only
  * with a limited warranty  and the software's author,  the holder of the
  * economic rights,  and the successive licensors  have only  limited
- * liability.
+ * liability. 
  *
  * In this respect, the user's attention is drawn to the risks associated
  * with loading,  using,  modifying and/or developing or reproducing the
@@ -22,39 +22,35 @@
  * therefore means  that it is reserved for developers  and  experienced
  * professionals having in-depth computer knowledge. Users are therefore
  * encouraged to load and test the software's suitability as regards their
- * requirements in conditions enabling the security of their systems and/or
- * data to be ensured and,  more generally, to use and operate it in the
- * same conditions as regards security.
+ * requirements in conditions enabling the security of their systems and/or 
+ * data to be ensured and,  more generally, to use and operate it in the 
+ * same conditions as regards security. 
  *
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-B license and that you accept its terms.
  */
-package fr.insalyon.creatis.vip.core.server.dao;
+package fr.insalyon.creatis.vip.publication.server.dao;
 
-import java.sql.Connection;
+import fr.insalyon.creatis.vip.core.server.dao.DAOException;
+import fr.insalyon.creatis.vip.publication.client.bean.Publication;
+
+import java.util.List;
 
 /**
- * @author Rafael Ferreira da Silva,Nouha boujelben
+ *
+ * @author Nouha Boujelben
  */
-public abstract class CoreDAOFactory {
+public interface PublicationDAO {
 
-    public static CoreDAOFactory getDAOFactory() {
+    public void add(Publication publication) throws DAOException;
 
-        return MySQLDAOFactory.getInstance();
-    }
+    public void update(Publication publication) throws DAOException;
 
-    public abstract UserDAO getUserDAO(Connection connection)
-        throws DAOException;
+    void updateOwnerEmail(String oldOwnerEmail, String newOwnerEmail) throws DAOException;
 
-    public abstract GroupDAO getGroupDAO(Connection connection)
-        throws DAOException;
+    public void remove(Long publicationID) throws DAOException;
+    
+    public Publication getPublication(Long publicationID) throws DAOException;
 
-    public abstract UsersGroupsDAO getUsersGroupsDAO(Connection connection)
-        throws DAOException;
-
-    public abstract AccountDAO getAccountDAO(Connection connection)
-        throws DAOException;
-
-    public abstract TermsUseDAO getTermsUseDAO(Connection connection)
-        throws DAOException;
+    public List<Publication> getList() throws DAOException;
 }
